add jwt and model dependent queue
add JSON Web Token as a user login service.
This makes it easier to make the proxy distributed and does not need any
user text file. And users can be added without restarting the server.
It also make it more secure in a non https env.

Extracted the queue part of the code to a subset of classes so different
queues can easily be implemented and used. As long as they conform to an
API.

Add a model depended queue, it prioritize the servers which have a
certain model already loaded. It also can have black list and white list of
which model it can handle due to cpu/ram/gpu. It conforms to the
shortest queue principle of the simple queue

Add logging instead of prints to the stderr output.

Add github-ci

Signed-off-by: Robert Marklund <robbelibobban@gmail.com>